Refactor FilterPosts
- extract FilterPostsMenuItems to its own component
- update tests accordingly
- add _.chunk from lodash to dry out code
